+++ CHANGELOG.md
@@ -4,12 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.9.0+100002] - 2022-12-27
 
 ### Added
 
-- Time remaining info for world map threats including settings option + localization.
+- New settings for timer/colors/glyphs + localization.
+- Time remaining info for world map threats.
 - Major Factions renown status in entry tooltip.
+
+### Changed
+
+- Updated details display in menu entry tooltips and restructured its code for better readability.
 
 ## [0.8.0+100002] - 2022-12-13